% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_Vg.R
\name{assemble_Ve}
\alias{assemble_Ve}
\title{Assemble into a matrix Ve from fit1_bvlmm() output}
\usage{
assemble_Ve(sigma)
}
\arguments{
\item{sigma}{Vector of output from fit1_bvlmm(); estimates of variance components}
}
\description{
Assemble into a matrix Ve from fit1_bvlmm() output
}
