% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_Vg.R
\name{assemble_Vg}
\alias{assemble_Vg}
\title{Assemble into a matrix Vg from fit1_bvlmm() output}
\usage{
assemble_Vg(sigma)
}
\arguments{
\item{sigma}{Vector of output from fit1_bvlmm(); estimates of variance components}
}
\description{
Assemble into a matrix Vg from fit1_bvlmm() output
}
