% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_det_rss.R
\name{calc_det_rss}
\alias{calc_det_rss}
\title{Calculate determinants of cross-product of residuals matrices}
\usage{
calc_det_rss(res)
}
\arguments{
\item{res}{an array of residual matrices}
}
\description{
Calculate determinants of cross-product of residuals matrices
}

