% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dets.R
\name{calc_dets}
\alias{calc_dets}
\title{Calculate determinant of 'squared residuals' at two loci for two (ordered) traits}
\usage{
calc_dets(yy, mat1, mat2)
}
\arguments{
\item{yy}{a n x 2 matrix containing two phenotypes}

\item{mat1}{a matrix of genotype (founder) probabilities}

\item{mat2}{a matrix of genotype (founder) probabilities}
}
\description{
Calculate determinant of 'squared residuals' at two loci for two (ordered) traits
}

