% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lik.R
\name{calc_lik}
\alias{calc_lik}
\title{Calculate likelihood at two loci for two (ordered) traits}
\usage{
calc_lik(yy, prob_mat1, prob_mat2)
}
\arguments{
\item{yy}{a n x 2 matrix containing two phenotypes}

\item{prob_mat1}{a matrix of genotype (founder) probabilities}

\item{prob_mat2}{a matrix of genotype (founder) probabilities}
}
\description{
Calculate likelihood at two loci for two (ordered) traits
}

