% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lod.R
\name{calc_lod}
\alias{calc_lod}
\title{Calculate LODs - two profile and one joint - for two traits}
\usage{
calc_lod(data, n_mouse, true_profile_lik = TRUE)
}
\arguments{
\item{data}{tibble containing log10determinants of RSS}

\item{n_mouse}{sample size}

\item{true_profile_lik}{indicator (defaults to TRUE) of whether to calculate true profile lod, ie, by maximizing over the nuisance parameter, qtl location.}
}
\description{
Calculate LODs - two profile and one joint - for two traits
}

