% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lod_list.R
\name{calc_lod_list}
\alias{calc_lod_list}
\title{Calculate LODs - two profile and one joint - for two traits}
\usage{
calc_lod_list(log10det_rss, n_mouse)
}
\arguments{
\item{log10det_rss}{matrix containing log10determinants of RSS}

\item{n_mouse}{sample size}
}
\description{
Calculate LODs - two profile and one joint - for two traits
}

