% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_profile_lod.R
\name{calc_profile_lod}
\alias{calc_profile_lod}
\title{Calculate a profile lod vector}
\usage{
calc_profile_lod(log10det_rss_mins, log10det_rss0, n_mouse)
}
\arguments{
\item{log10det_rss_mins}{a vector of log10 determinant minima by row or by column}

\item{log10det_rss0}{a vector of log10 determinant under pleiotropy at distinct loci}

\item{n_mouse}{number of mice for likelihood calculations}
}
\description{
Calculate a profile lod vector
}

