% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_profile_lods.R
\name{calc_profile_lods}
\alias{calc_profile_lods}
\title{Calculate profile LODs for two traits}
\usage{
calc_profile_lods(log10det_rss)
}
\arguments{
\item{log10det_rss}{matrix containing log10determinants of RSS}
}
\description{
Calculate profile LODs for two traits
}

