% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_res.R
\name{calc_res}
\alias{calc_res}
\title{Calculate an array of residuals using Haley-Knott regression}
\usage{
calc_res(yy, geno_array)
}
\arguments{
\item{yy}{n x 2 data matrix}

\item{geno_array}{genotype probabilities array}
}
\description{
Calculate an array of residuals using Haley-Knott regression
}
