% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rss.R
\name{calc_rss_one}
\alias{calc_rss_one}
\title{Calculate RSS for a single locus and a single trait}
\usage{
calc_rss_one(y, genomat)
}
\arguments{
\item{y}{a vector of phenotype values, length n}

\item{genomat}{a single genotype matrix, n x 8}
}
\description{
Calculate RSS for a single locus and a single trait
}

