% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit1_mvlmm.R
\name{fit1_bvlmm}
\alias{fit1_bvlmm}
\title{Fit a 2-locus, multivariate trait linear mixed effects model}
\usage{
fit1_bvlmm(Y, X1, X2, Kmat, reml = TRUE)
}
\arguments{
\item{Y}{a phenotypes matrix with one row per subject and one trait per column}

\item{X1}{genotype probabilities for first locus, one row per subject}

\item{X2}{genotype probabilities for second locus, one row per subject}

\item{Kmat}{a kinship matrix}

\item{reml}{indicator of whether to use REML method}
}
\description{
Fit a 2-locus, multivariate trait linear mixed effects model
}
