% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_residuals.R
\name{hk_residuals}
\alias{hk_residuals}
\title{Calculate residuals with Haley-Knott regression}
\usage{
hk_residuals(pheno, genomat)
}
\arguments{
\item{pheno}{a phenotype vector or matrix (with each column being one phenotype)}

\item{genomat}{a matrix of genotype probabilities, with the first column of probabilities replaced with a column of 1's.}
}
\description{
Calculate residuals with Haley-Knott regression
}

