% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_pheno.R
\name{prep_pheno}
\alias{prep_pheno}
\title{Prepare phenotype for use with scan functions}
\usage{
prep_pheno(phenotype, alpha = 0.02, transform = log)
}
\arguments{
\item{phenotype}{a phenotype vector with values for a single trait}

\item{alpha}{input to winsorize}

\item{transform}{function to apply to phenotype}
}
\description{
Prepare phenotype for use with scan functions
}

