% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_one.R
\name{scan_one}
\alias{scan_one}
\title{One-dimensional scan over a region of interest using only intercept and genotypes}
\usage{
scan_one(y, genoarray, start, stop)
}
\arguments{
\item{y}{a vector of n phenotype values}

\item{genoarray}{a three-dimensional array of genotypes, like `probs$probs$`1``}

\item{start}{index for start of scan}

\item{stop}{index for stop of scan}
}
\description{
One-dimensional scan over a region of interest using only intercept and genotypes
}

