% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pheno.R
\name{sim_pheno}
\alias{sim_pheno}
\title{Simulate phenotypes from genotype data}
\usage{
sim_pheno(genomat, beta = 1:8/8, sd = 0.05)
}
\arguments{
\item{genomat}{a matrix of genotype probabilities}

\item{beta}{a vector (length 8) of genotype class effects}

\item{sd}{standard deviation of simulated noise}
}
\description{
Simulate phenotypes from genotype data
}
