% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pheno2.R
\name{sim_pheno2}
\alias{sim_pheno2}
\title{Simulate two phenotypes simultaneously}
\usage{
sim_pheno2(genomat, bmat, Sigma)
}
\arguments{
\item{genomat}{a matrix of genotypes}

\item{bmat}{a matrix of effect sizes}

\item{Sigma}{a 2x2 covariance matrix}
}
\description{
Simulate two phenotypes simultaneously
}
