% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pheno_lmm.R
\name{sim_pheno_lmm}
\alias{sim_pheno_lmm}
\title{Simulate bivariate phenotype with incorporation of random effects}
\usage{
sim_pheno_lmm(genomat, bmat, Sigma)
}
\arguments{
\item{genomat}{a single locus's genotypes probability matrix}

\item{bmat}{a matrix of effect sizes}

\item{Sigma}{a 2n by 2n covariance matrix}
}
\description{
Simulate bivariate phenotype with incorporation of random effects
}
\examples{
Kmat <- diag(1, 8)
Vg <- matrix(c(1:4 / 10), nrow = 2, ncol = 2)
Ve <- matrix(c(2:5) / 10, nrow = 2, ncol = 2)
In <- diag(1, 8)
sim_pheno_lmm(genomat = matrix(c(1, 0, 0, 1, 2, 1, 0, 1), ncol = 1), 
bmat = matrix(c(2.5, 3), nrow = 1), 
Sigma = Vg \%x\% Kmat + Ve \%x\% In)

}
