% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ll_bvlmm.R
\name{calc_ll_bvlmm}
\alias{calc_ll_bvlmm}
\title{Calculate log-likelihood for a LMM}
\usage{
calc_ll_bvlmm(yvec, Xmat, betahat, var)
}
\arguments{
\item{yvec}{a vector of phenotype values}

\item{Xmat}{a design matrix}

\item{betahat}{estimate of beta}

\item{var}{a covariance matrix}
}
\description{
Calculate log-likelihood for a LMM
}
