% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lod.R
\name{calc_lod}
\alias{calc_lod}
\title{Calculate LODs - two profile and one joint - for two traits}
\usage{
calc_lod(data, n_mouse)
}
\arguments{
\item{data}{tibble containing log10determinants of RSS}

\item{n_mouse}{sample size}
}
\description{
Calculate LODs - two profile and one joint - for two traits
}
