% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_res.R
\name{calc_res}
\alias{calc_res}
\title{Calculate an array of residuals using Haley-Knott regression}
\usage{
calc_res(yy, geno_array, start, stop)
}
\arguments{
\item{yy}{n x 2 data matrix}

\item{geno_array}{genotype probabilities array}

\item{start}{marker index for scan start}

\item{stop}{marker index for scan stop}
}
\description{
Calculate an array of residuals using Haley-Knott regression
}

