% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rss.R
\name{calc_rss}
\alias{calc_rss}
\title{One-dimensional scan over a region of interest using only intercept and genotypes}
\usage{
calc_rss(y, genoarray)
}
\arguments{
\item{y}{a vector of n phenotype values}

\item{genoarray}{a three-dimensional array of genotypes, like `probs$probs$`1``}
}
\description{
One-dimensional scan over a region of interest using only intercept and genotypes
}
