% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rss.R
\name{calc_rss}
\alias{calc_rss}
\title{Calculate determinant of 'squared residuals' (ie, RSS) at two loci for two (ordered) traits}
\usage{
calc_rss(yy, mat1, mat2)
}
\arguments{
\item{yy}{a n x 2 matrix containing two phenotypes}

\item{mat1}{a matrix of genotype (founder) probabilities with first column being all 1's}

\item{mat2}{a matrix of genotype (founder) probabilities with first column being all 1's}
}
\description{
Calculate determinant of 'squared residuals' (ie, RSS) at two loci for two (ordered) traits
}

