% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{profile_plot}
\alias{profile_plot}
\title{Plot profile LR statistics for two traits}
\usage{
profile_plot(profile1, profile2, marker_positions, col1 = "red",
  col2 = "blue")
}
\arguments{
\item{profile1}{profile lr statistic for trait one}

\item{profile2}{profile lr statistic for trait two}

\item{marker_positions}{vector of marker positions (should have length equal to that of profile1 & profile2)}

\item{col1}{color of trait 1 plot}

\item{col2}{color of trait 2 plot}
}
\description{
Plot profile LR statistics for two traits
}

